scraper.py
#!/usr/bin/env python3
import requests
from bs4 import BeautifulSoup
import json
import os
import yaml
import time
import random
from datetime import datetime

# User agent to mimic a browser
USER_AGENT = "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36"

# List of deal websites to scrape (using placeholder URLs for demonstration)
# In a real implementation, you would use actual deal websites
DEAL_SOURCES = [
    {
        "name": "Example Deals",
        "url": "https://example.com/deals",
        "item_selector": ".deal-item",
        "title_selector": ".title",
        "price_selector": ".price",
        "discount_selector": ".discount",
        "link_selector": "a",
        "image_selector": "img"
    },
    {
        "name": "Sample Shop",
        "url": "https://sample-shop.com/offers",
        "item_selector": ".product-card",
        "title_selector": "h3",
        "price_selector": ".current-price",
        "discount_selector": ".discount-badge",
        "link_selector": "a",
        "image_selector": ".product-image"
    }
]

def scrape_deals():
    """Scrape deals from all sources and return a combined list."""
    all_deals = []
    
    for source in DEAL_SOURCES:
        try:
            print(f"Scraping deals from {source['name']}...")
            
            # In a real implementation, this would make an actual HTTP request
            # For demonstration, we'll simulate the response
            
            # Simulated data for demonstration purposes
            simulated_deals = [
                {
                    "title": f"Alien Blaster 3000 - {source['name']}",
                    "price": "$49.99",
                    "original_price": "$99.99",
                    "discount": "50%",
                    "link": "https://example.com/product/alien-blaster",
                    "image": "https://example.com/images/alien-blaster.jpg",
                    "source": source['name'],
                    "date_added": datetime.now().strftime("%Y-%m-%d %H:%M:%S")
                },
                {
                    "title": f"UFO Detection Kit - {source['name']}",
                    "price": "$29.99",
                    "original_price": "$59.99",
                    "discount": "50%",
                    "link": "https://example.com/product/ufo-detector",
                    "image": "https://example.com/images/ufo-detector.jpg",
                    "source": source['name'],
                    "date_added": datetime.now().strftime("%Y-%m-%d %H:%M:%S")
                },
                {
                    "title": f"Martian Translator - {source['name']}",
                    "price": "$19.99",
                    "original_price": "$39.99",
                    "discount": "50%",
                    "link": "https://example.com/product/martian-translator",
                    "image": "https://example.com/images/martian-translator.jpg",
                    "source": source['name'],
                    "date_added": datetime.now().strftime("%Y-%m-%d %H:%M:%S")
                }
            ]
            
            all_deals.extend(simulated_deals)
            
            # Add a delay to avoid overloading servers (important for real implementation)
            time.sleep(random.uniform(1, 3))
            
        except Exception as e:
            print(f"Error scraping {source['name']}: {e}")
    
    return all_deals

def save_deals_json(deals):
    """Save deals to a JSON file."""
    with open("deals.json", "w") as f:
        json.dump(deals, f, indent=2)
    print(f"Saved {len(deals)} deals to deals.json")

def save_deals_yaml(deals):
    """Save deals to a YAML file for Jekyll."""
    # Create _data directory if it doesn't exist
    os.makedirs("_data", exist_ok=True)
    
    with open("_data/deals.yml", "w") as f:
        yaml.dump(deals, f, sort_keys=False)
    print(f"Saved {len(deals)} deals to _data/deals.yml")

def main():
    """Main function to run the scraper."""
    print("Starting deal scraper...")
    deals = scrape_deals()
    save_deals_json(deals)
    save_deals_yaml(deals)
    print("Deal scraping completed!")

if __name__ == "__main__":
    main() 