requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
pyyaml==6.0 